% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_orfs.R
\name{.build_granges}
\alias{.build_granges}
\title{Build a GRanges from raw data and a spec}
\usage{
.build_granges(raw, spec, additional_cols = NULL)
}
\description{
Build a GRanges from raw data and a spec
}
\keyword{internal}
