% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_orfs.R
\name{.extend_ribocode_stop_codon}
\alias{.extend_ribocode_stop_codon}
\title{Extend RiboCode ORF coordinates to include stop codon (simple genomic extension)}
\usage{
.extend_ribocode_stop_codon(gr, raw_df, txdb)
}
\arguments{
\item{gr}{GRanges of ORFs}

\item{raw_df}{Raw RiboCode data.frame (not currently used)}

\item{txdb}{TxDb object (not used in simple version, kept for compatibility)}
}
\value{
GRanges with extended coordinates
}
\description{
RiboCode reports genomic coordinates excluding the stop codon.
This function extends each ORF by 3 nucleotides in genomic space
to include the stop codon (fast, but doesn't account for splicing).
}
\keyword{internal}
