% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_orfs.R
\name{.make_full_output}
\alias{.make_full_output}
\title{Create full output (list with unified GRangesList and metadata table)}
\usage{
.make_full_output(gr, raw_data)
}
\description{
Create full output (list with unified GRangesList and metadata table)
}
\keyword{internal}
