% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_orfs.R
\name{.make_unified_output}
\alias{.make_unified_output}
\title{Create unified output (GRanges with core metadata)}
\usage{
.make_unified_output(gr, additional_cols = NULL)
}
\description{
Create unified output (GRanges with core metadata)
}
\keyword{internal}
