% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_orfs.R
\name{.mcols_to_dataframe}
\alias{.mcols_to_dataframe}
\title{Convert GRanges mcols to a simple data.frame
Handles complex column types like GRanges, AAStringSet, CompressedLists}
\usage{
.mcols_to_dataframe(gr)
}
\description{
Convert GRanges mcols to a simple data.frame
Handles complex column types like GRanges, AAStringSet, CompressedLists
}
\keyword{internal}
