% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reannotate_orf_type.R
\name{orf_type_levels}
\alias{orf_type_levels}
\title{Supported reference ORF types}
\usage{
orf_type_levels()
}
\value{
A character vector of supported ORF types.
}
\description{
Return the ordered vocabulary used by \code{\link[=reannotate_orf_type]{reannotate_orf_type()}}.
}
