% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_orfs_isoforms.R
\name{annotate_orf_isoforms}
\alias{annotate_orf_isoforms}
\title{Annotate ORF with their status respective of}
\usage{
annotate_orf_isoforms(
  annotations,
  orfs,
  BSgenome,
  transcript_meta,
  orfs_meta = NULL,
  start_codons = c("ATG", "TTG", "CTG", "GTG"),
  stop_codons = c("TAG", "TAA", "TGA")
)
}
\arguments{
\item{annotations}{List with genomic annotations, output of \code{prepare_annotations()}}

\item{orfs}{GRanges or GRangesList with genomic coordinates of ORFs of interest}

\item{transcript_meta}{(optional) A data.frame with transcript metadata (for example: gene_name, gene_id, transcript_type), must contain a column \code{transcript_id} compatible with \code{names(annotations$transcripts)}}

\item{orfs_meta}{(optional) A data.frame with ORF metadata, must contain a column \code{ORF_id} compatible with \code{names(orfs)}}

\item{start_codons}{A vector with codons recognized as start codon; default \code{c("ATG","TTG","CTG","GTG")}}

\item{stop_codons}{A vector with codons recognized as stop codon; default \code{c("TAG", "TAA", "TGA")}}
}
\value{
A list containing a table with annotated ORF-transcript pairs (\verb{$table}) and GRangesList with genomic coordinates of identified ORFs (\verb{$ranges})
}
\description{
Annotate ORF with their status respective of
}
\examples{
BSgenome <- BSgenome.Hsapiens.UCSC.hg38
gtf <- "inst/extdata/gencode.v35.annotation_chr10.gtf"
bed <- "inst/extdata/Ribo-seq_ORFs.bed"

# Prepare annotations
annotations <- prepare_annotations_fromGTF(gtf, BSgenome)

# Pull transcripts metadata
transcripts_meta <- import(gtf, format = "GTF") \%>\%
  as.data.frame() \%>\%
  dplyr::filter(type == "transcript") \%>\%
  dplyr::select(gene_name, transcript_id, transcript_type) \%>\%
  distinct()

# Load ORFs
orfs <- import(bed, format = "BED")
names(orfs) <- orfs$name
orf_tab <- as.data.frame(orfs)
annotated_orfs <- annotate_orf_isoforms(annotations, orfs, BSgenome, transcripts_meta)
}
