% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_seq_levels}
\alias{check_seq_levels}
\title{Title}
\usage{
check_seq_levels(x, y)
}
\arguments{
\item{x, y}{The object from/on which to get/set the seqlevels and seqlevels style}
}
\description{
Title
}
\examples{
check_seq_levels(Granges_1, Granges_2)
}
\keyword{internal}
