% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_orf_usage.R
\name{diff_orf_usage}
\alias{diff_orf_usage}
\title{Calculate Differential ORF Usage}
\usage{
diff_orf_usage(
  annotated_orfs_tab,
  dtu_results,
  selected_genes,
  orf_translation = NULL
)
}
\arguments{
\item{annotated_orfs_tab}{A data frame containing information about annotated ORFs.}

\item{dtu_results}{A data frame containing results of differential transcript usage analysis.}

\item{selected_genes}{A character vector of gene IDs for which differential ORF usage will be calculated.}

\item{orf_translation}{(optional) A data frame containing ORF translations.}
}
\value{
A list containing two elements: 'gene_stats' and 'gene_tables'.
\itemize{
\item 'gene_stats': A data frame with statistics about the differential ORF usage for each gene.
\item 'gene_tables': A list of data tables, each containing the ORF/isoform combinations for a specific gene.
}
}
\description{
This function calculates differential ORF usage based on the provided inputs.
}
\examples{
# Example Usage
diff_orf_usage(annotated_orfs_tab, dtu_results, selected_genes)

}
