% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations.R
\name{prepare_annotations_fromGTF}
\alias{prepare_annotations_fromGTF}
\title{Prepare annotation set}
\usage{
prepare_annotations_fromGTF(gtf, BSgenome)
}
\arguments{
\item{gtf}{Path to a GTF file including transcript isoform models}

\item{BSgenome}{BSgenome object compatible with provided GTF file}
}
\value{
GRangesList with genomic coordinates of different genomic regions
}
\description{
Prepare annotation set
}
\examples{

library(BSgenome.Hsapiens.UCSC.hg38) # Load selected BSgenome
gtf <- "gencode.v29.annotations.gtf"
annotations <- prepare_annotations_fromGTF(gtf, BSgenome.Hsapiens.UCSC.hg38)
}
