% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_exons}
\alias{rank_exons}
\title{Function to assign appropriate exon ranking depending on the ranges strand}
\usage{
rank_exons(grl)
}
\arguments{
\item{grl}{GRangesList object}
}
\value{
A list of genomic ranges objects with exon ranks assigned.
}
\description{
Function to assign appropriate exon ranking depending on the ranges strand
}
